// test-aggrid-big/front-ng18/src/app/generated/rest-bal/model/firstnameDTO.ts
/**
 * 
 * 
 *
 * 
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export interface FirstnameDTO { 
    prenom?: string;
    genre?: string;
    langage?: string;
    frequence?: number;
}